Take repository URL from the commandline.
To make usage of twin within a CI setup easier, make it possible to
run twin without a pypirc file. This commit adds the possibility to
specify the repository URL directly on the commandline.

This is done by overloading the `--repository` switch: this can now be
a section in the pypirc file (as before) or a full URL (e.g. something
with `://` in it).

Note that this is only needed when you don't want to upload to PyPI,
for PyPI we already have a fallback in place.